+++ auction-service/src/main/java/com/bugzero/rarego/boundedContext/auction/app/AuctionRelistUseCase.java
@@ -12,8 +12,8 @@ import com.bugzero.rarego.boundedContext.auction.out.AuctionOrderRepository;
 import com.bugzero.rarego.boundedContext.auction.out.AuctionRepository;
 import com.bugzero.rarego.global.exception.CustomException;
 import com.bugzero.rarego.global.response.ErrorType;
-import com.bugzero.rarego.shared.auction.dto.AuctionRelistRequestDto;
-import com.bugzero.rarego.shared.auction.dto.AuctionRelistResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionRelistRequestDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionRelistResponseDto;
 
 import lombok.RequiredArgsConstructor;
