+++ auction-service/src/main/java/com/bugzero/rarego/boundedContext/auction/in/AuctionController.java
@@ -2,13 +2,22 @@ package com.bugzero.rarego.boundedContext.auction.in;
 
 import com.bugzero.rarego.boundedContext.auction.app.AuctionFacade;
 import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionAddBookmarkResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionDetailResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionListResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionOrderResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionRelistRequestDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionRelistResponseDto;
 import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionRemoveBookmarkResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionSearchCondition;
 import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionWithdrawResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.BidLogResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.BidRequestDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.BidResponseDto;
 import com.bugzero.rarego.global.response.PagedResponseDto;
 import com.bugzero.rarego.global.response.SuccessResponseDto;
 import com.bugzero.rarego.global.response.SuccessType;
 import com.bugzero.rarego.global.security.MemberPrincipal;
-import com.bugzero.rarego.shared.auction.dto.*;
+
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
