+++ auction-service/src/main/java/com/bugzero/rarego/boundedContext/auction/in/AuctionMemberController.java
@@ -6,10 +6,10 @@ import com.bugzero.rarego.boundedContext.auction.domain.AuctionStatus;
 import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionBookmarkListResponseDto;
 import com.bugzero.rarego.global.response.PagedResponseDto;
 import com.bugzero.rarego.global.security.MemberPrincipal;
-import com.bugzero.rarego.shared.auction.dto.AuctionFilterType;
-import com.bugzero.rarego.shared.auction.dto.MyAuctionOrderListResponseDto;
-import com.bugzero.rarego.shared.auction.dto.MyBidResponseDto;
-import com.bugzero.rarego.shared.auction.dto.MySaleResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionFilterType;
+import com.bugzero.rarego.boundedContext.auction.in.dto.MyAuctionOrderListResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.MyBidResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.MySaleResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
