+++ auction-service/src/main/java/com/bugzero/rarego/boundedContext/auction/in/dto/BidResponseDto.java
@@ -1,4 +1,4 @@
-package com.bugzero.rarego.shared.auction.dto;
+package com.bugzero.rarego.boundedContext.auction.in.dto;
 
 import com.bugzero.rarego.boundedContext.auction.domain.Bid;
 import java.time.LocalDateTime;
