+++ auction-service/src/main/java/com/bugzero/rarego/boundedContext/auction/in/dto/MyAuctionOrderListResponseDto.java
@@ -1,4 +1,4 @@
-package com.bugzero.rarego.shared.auction.dto;
+package com.bugzero.rarego.boundedContext.auction.in.dto;
 
 import java.time.LocalDateTime;
