+++ auction-service/src/test/java/com/bugzero/rarego/boundedContext/auction/in/AuctionControllerTest.java
@@ -4,12 +4,18 @@ import com.bugzero.rarego.boundedContext.auction.app.AuctionFacade;
 import com.bugzero.rarego.boundedContext.auction.domain.AuctionOrderStatus;
 import com.bugzero.rarego.boundedContext.auction.domain.AuctionStatus;
 import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionAddBookmarkResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionDetailResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionOrderResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionRelistRequestDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionRelistResponseDto;
 import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionRemoveBookmarkResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.BidLogResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.BidRequestDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.BidResponseDto;
 import com.bugzero.rarego.global.exception.CustomException;
 import com.bugzero.rarego.global.exception.GlobalExceptionHandler;
 import com.bugzero.rarego.global.response.*;
 import com.bugzero.rarego.global.security.MemberPrincipal;
-import com.bugzero.rarego.shared.auction.dto.*;
 import com.bugzero.rarego.support.WithMockMemberPrincipal;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
