+++ src/test/java/com/bugzero/rarego/boundedContext/auction/app/AuctionCreateBidUseCaseTest.java
@@ -8,7 +8,7 @@ import com.bugzero.rarego.boundedContext.auction.event.AuctionBidCreatedEvent;
 import com.bugzero.rarego.boundedContext.auction.out.BidRepository;
 import com.bugzero.rarego.global.exception.CustomException;
 import com.bugzero.rarego.global.response.ErrorType;
-import com.bugzero.rarego.shared.auction.dto.BidResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.BidResponseDto;
 import com.bugzero.rarego.shared.payment.out.PaymentApiClient;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
