+++ src/test/java/com/bugzero/rarego/boundedContext/auction/in/AuctionMemberControllerTest.java
@@ -10,9 +10,9 @@ import com.bugzero.rarego.global.response.ErrorType;
 import com.bugzero.rarego.global.response.PageDto;
 import com.bugzero.rarego.global.response.PagedResponseDto;
 import com.bugzero.rarego.global.security.MemberPrincipal;
-import com.bugzero.rarego.shared.auction.dto.AuctionListResponseDto;
-import com.bugzero.rarego.shared.auction.dto.MyAuctionOrderListResponseDto;
-import com.bugzero.rarego.shared.auction.dto.MyBidResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.AuctionListResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.MyAuctionOrderListResponseDto;
+import com.bugzero.rarego.boundedContext.auction.in.dto.MyBidResponseDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
